+++ packages/components-react/rollup.config.js
@@ -1,30 +1,27 @@
 import resolve from '@rollup/plugin-node-resolve';
-import replace from '@rollup/plugin-replace';
 import polyfill from 'rollup-plugin-polyfill';
 import pkg from '@porsche-design-system/js/package.json';
 
+const polyfills = [
+  'construct-style-sheets-polyfill',
+  'intersection-observer',
+  'scroll-behavior-polyfill',
+  'isomorphic-fetch',
+];
+
 export default {
-  input: 'projects/components-wrapper/src/bundle/index.js',
+  input: 'projects/jsdom-polyfill/src/index.js',
   output: {
     esModule: false,
     dir: 'dist/components-wrapper/jsdom-polyfill',
-    format: 'umd',
+    format: 'cjs',
     name: pkg.name,
     extend: true,
   },
   plugins: [
-    polyfill([
-      'construct-style-sheets-polyfill',
-      'intersection-observer',
-      'scroll-behavior-polyfill',
-      'isomorphic-fetch'
-    ]),
+    polyfill(polyfills),
     resolve({
-      resolveOnly: [/^@stencil\/.*$/, /^@porsche-design-system\/.*$/],
-    }),
-    replace({
-      // inject call of 'defineCustomElements()'
-      'exports.PBanner': 'defineCustomElements();\n\n  exports.PBanner',
+      resolveOnly: polyfills,
     }),
   ],
 };