+++ packages/components-react/tests/jsdom-polyfill/specs/banner.spec.tsx
@@ -1,7 +1,5 @@
 import React, { useState } from 'react';
-import {
-  componentsReady, PBanner
-} from '@porsche-design-system/components-react';
+import { componentsReady, PBanner } from '@porsche-design-system/components-react';
 import { render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
@@ -11,14 +9,19 @@ const Sample = (): JSX.Element => {
 
   return (
     <>
-      <PBanner data-testid="host" onDismiss={(e) => {
-        setEventCounter(eventCounter + 1);
-        setStatus('hidden');
-      }}>
+      <PBanner
+        data-testid="host"
+        onDismiss={(e) => {
+          setEventCounter(eventCounter + 1);
+          setStatus('hidden');
+        }}
+      >
         <span slot="title">Some banner title</span>
         <span slot="description">Some banner description.</span>
       </PBanner>
-      <div data-testid="debug">{`Status: ${status};`} {`Event Counter: ${eventCounter};`}</div>
+      <div data-testid="debug">
+        {`Status: ${status};`} {`Event Counter: ${eventCounter};`}
+      </div>
     </>
   );
 };
@@ -26,7 +29,6 @@ const Sample = (): JSX.Element => {
 describe('PBanner', () => {
   it('should have initialized shadow dom', async () => {
     const { getByTestId } = render(<Sample />);
-
     await componentsReady();
 
     expect(getByTestId('host').shadowRoot).not.toBeNull();
@@ -34,7 +36,6 @@ describe('PBanner', () => {
 
   it('should have working events', async () => {
     const { getByTestId } = render(<Sample />);
-
     await componentsReady();
 
     const debug = getByTestId('debug');