+++ packages/components-react/tests/jsdom-polyfill/specs/radio-button.spec.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
-import {
-  componentsReady, PRadioButtonWrapper
-} from '@porsche-design-system/components-react';
+import { componentsReady, PRadioButtonWrapper } from '@porsche-design-system/components-react';
 import { render } from '@testing-library/react';
 
 const Sample = (): JSX.Element => {
@@ -17,7 +15,6 @@ const Sample = (): JSX.Element => {
 describe('PRadioButtonWrapper', () => {
   it('should have initialized shadow dom', async () => {
     const { getByTestId } = render(<Sample />);
-
     await componentsReady();
 
     expect(getByTestId('host').shadowRoot).not.toBeNull();