+++ packages/components-react/tests/jsdom-polyfill/specs/text.spec.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
-import {
-  componentsReady, PText
-} from '@porsche-design-system/components-react';
+import { componentsReady, PText } from '@porsche-design-system/components-react';
 import { render } from '@testing-library/react';
 
 const Sample = (): JSX.Element => {
@@ -15,7 +13,6 @@ const Sample = (): JSX.Element => {
 describe('PText', () => {
   it('should have initialized shadow dom', async () => {
     const { getByTestId } = render(<Sample />);
-
     await componentsReady();
 
     expect(getByTestId('host').shadowRoot).not.toBeNull();